Improved sidebar clicks and sign out functionality
